requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
astor==0.8.1
attrs==23.1.0
blinker==1.7.0
cassandra-driver==3.28.0
cassio==0.1.3
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.3
distro==1.8.0
Flask==3.0.0
frozenlist==1.4.0
geomet==0.2.1.post1
graphviz==0.20.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.345
langchain-core==0.0.9
langsmith==0.0.69
MarkupSafe==2.1.3
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.2
openai==1.3.7
packaging==23.2
parso==0.8.3
Pillow==10.1.0
pluggy==1.3.0
pudb==2023.1
py2cfg==0.7.3
pydantic==2.5.2
pydantic_core==2.14.5
Pygments==2.17.2
pytest==7.4.3
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
tenacity==8.2.3
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==2.1.0
urwid==2.2.3
urwid-readline==0.13
websocket-server==0.6.4
Werkzeug==3.0.1
yarl==1.9.3
